Fix reading upper 8 bits in spi_transaction
Previously, the uppermost byte of the three is read, but when shifting
it into the spi_transaction result value, it would overflow the 16 bits
used for this expression (due to integer promotion). This meant that
this byte would effectively always read as 0, which might not be a
problem in how spi_transaction was used so far.

+++ code.cpp
@@ -475,6 +475,6 @@ uint16_t spi_transaction (uint8_t a, uint8_t b, uint8_t c, uint8_t d) {
   Serial.print(" 0x"); Serial.print(m, HEX);
   Serial.print(" 0x"); Serial.println(r, HEX);
   */
-  return 0xFFFFFF & ((n<<16)+(m<<8) + r);
+  return 0xFFFFFF & (((uint32_t)n<<16)+(m<<8) + r);
 }
